+++ src/lang/index.ts
@@ -11,6 +11,7 @@ const i18n = new VueI18n({
   locale: "en",
   fallbackLocale: "en",
   messages: { en, de, fr },
+  silentTranslationWarn: process.env.NODE_ENV === "production",
 });
 
 export default i18n;