+++ vue.config.js
@@ -1,4 +1,4 @@
-const MomentLocalesPlugin = require("moment-locales-webpack-plugin");
+const { ContextReplacementPlugin } = require("webpack");
 const { version, author, homepage, bugs } = require("./package.json");
 
 process.env.VUE_APP_TITLE = "Radiolise";
@@ -7,15 +7,18 @@ process.env.VUE_APP_COPYRIGHT = `© 2017-${new Date().getFullYear()} ${author}`;
 process.env.VUE_APP_REPO = homepage;
 process.env.VUE_APP_ISSUES = bugs.url;
 
+const dateFnsLocales = ["de", "fr" /*, "be", "ru" */];
+
 module.exports = {
   publicPath: ".",
   productionSourceMap: false,
 
   configureWebpack: config => {
     config.plugins.push(
-      new MomentLocalesPlugin({
-        localesToKeep: ["en", "de", "fr" /*, "be", "ru" */],
-      })
+      new ContextReplacementPlugin(
+        /date\-fns[\/\\]/,
+        new RegExp(`[/\\\\\](${dateFnsLocales.join("|")})[/\\\\\]index\.js$`)
+      )
     );
   },
 
@@ -23,6 +26,12 @@ module.exports = {
     config.optimization.minimizer("terser").tap(args => {
       args[0].terserOptions.keep_classnames = true;
       return args;
+    });
+
+    config.plugin("prefetch").tap(options => {
+      options[0].fileBlacklist = options[0].fileBlacklist || [];
+      options[0].fileBlacklist.push(/date-fns-locale-(.)+?\.js$/);
+      return options;
     });
   },
 };